re-organize the folder
