+++ frontend/src/components/Button/NavButton/NavButton.css
@@ -1,0 +1,18 @@
+.nav-button {
+    font-family: "Tiro Devanagari Hindi";
+    text-decoration: none;
+    color: #FFFFFF;
+    white-space: nowrap;
+    font-size: clamp(0.9rem, 3vw, 1.5rem);
+    padding: 5px;
+    padding-left: clamp(15px, 3vw, 20px);
+    padding-right: clamp(15px, 3vw, 20px);
+    background-color: rgb(122, 122, 122);  
+    border: none;
+    border-radius: 10px;
+    box-shadow: 0px 4px 4px rgba(0,0,0,0.25);
+    transition-duration: 0.4s;
+    &:hover {
+        background-color: rgb(64, 64, 64);
+    }
+}
