+++ frontend/src/components/Functions/TopNavBar/TopNavBar.css
@@ -7,8 +7,7 @@
     top: 0; 
     left: 0;
     width: 100%;
-    min-height: 60px;
-    /* make topnavbar on top of other content */
+    min-height: 40px;
     z-index: 3;
 }
   
@@ -48,25 +47,6 @@
     padding: 0;
     margin: 0;
     gap: clamp(1rem, 3vw, 2rem);
-}
-
-.nav-button {
-    font-family: "Tiro Devanagari Hindi";
-    text-decoration: none;
-    color: #FFFFFF;
-    white-space: nowrap;
-    font-size: clamp(0.9rem, 3vw, 1.5rem);
-    padding: 5px;
-    padding-left: clamp(15px, 3vw, 20px);
-    padding-right: clamp(15px, 3vw, 20px);
-    background-color: rgb(122, 122, 122);  
-    border: none;
-    border-radius: 10px;
-    box-shadow: 0px 4px 4px rgba(0,0,0,0.25);
-    transition-duration: 0.4s;
-    &:hover {
-        background-color: rgb(64, 64, 64);
-    }
 }
 
 .userIcon{
